+++ src/argilla/server/daos/models/datasets.py
@@ -15,8 +15,8 @@
 from datetime import datetime
 from typing import Any, Dict, Optional, TypeVar, Union
 
-from argilla._constants import ES_INDEX_REGEX_PATTERN
 from argilla.server.commons.models import TaskType
+from argilla.server.constants import ES_INDEX_REGEX_PATTERN
 from argilla.server.pydantic_v1 import BaseModel, Field, root_validator