+++ src/argilla/server/security/model.py
@@ -16,7 +16,7 @@ from datetime import datetime
 from typing import Any, List, Optional
 from uuid import UUID
 
-from argilla._constants import ES_INDEX_REGEX_PATTERN
+from argilla.server.constants import ES_INDEX_REGEX_PATTERN
 from argilla.server.models import UserRole
 from argilla.server.pydantic_v1 import BaseModel, Field, constr
 from argilla.server.pydantic_v1.utils import GetterDict