+++ tests/unit/server/api/v0/test_metrics.py
@@ -15,7 +15,6 @@
 from typing import TYPE_CHECKING
 
 import pytest
-from argilla._constants import API_KEY_HEADER_NAME
 from argilla.server.apis.v0.models.text2text import (
     Text2TextBulkRequest,
     Text2TextRecord,
@@ -29,6 +28,7 @@ from argilla.server.apis.v0.models.token_classification import (
     TokenClassificationRecord,
 )
 from argilla.server.commons.models import TaskType
+from argilla.server.constants import API_KEY_HEADER_NAME
 from argilla.server.models import User, UserRole
 from argilla.server.services.metrics.models import CommonTasksMetrics
 from argilla.server.services.tasks.text_classification.metrics import TextClassificationMetrics